% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{data.frameORSummarizedExperiment-class}
\alias{data.frameORSummarizedExperiment-class}
\title{S4 Class union}
\description{
Class union allowing \code{MarrData} slot to be a data.frame or
Summarized Experiment
}
