% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{marr-class}
\alias{marr-class}
\title{the marr class}
\value{
\code{marrSamplepairs} returns the distribution of
percent reproducible  features (column-wise) per sample pair,
\code{marrFeatures} returns the  distribution of percent reproducible
sample pairs (row-wise) per feature,
\code{marrSamplepairsfiltered} returns the percent of reproducible
features based on a threshold value and
\code{marrFeaturesfiltered} returns the percent of reproducible
sample pairs based on a threshold value
}
\description{
Objects of this class store
needed information to work with a
marr object
}
\section{Slots}{

\describe{
\item{\code{marrSamplepairs}}{marr sample pairs}

\item{\code{marrFeatures}}{marr features}

\item{\code{marrSamplepairsfiltered}}{marr sample pairs post filtering}

\item{\code{marrFeaturesfiltered}}{marr metabolites post filtering}
}}

\examples{
data <- matrix(rnorm(2400), nrow=200, ncol=12)
data_marr <- marr(object = data, pSamplepairs=0.75,
pFeatures=0.75, alpha=0.05)

}
