% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllslotAccessors.R
\docType{methods}
\name{marrFeaturesfiltered}
\alias{marrFeaturesfiltered}
\alias{marrFeaturesfiltered,marr-method}
\title{Generic function that returns the marr filtered features}
\usage{
marrFeaturesfiltered(object)

\S4method{marrFeaturesfiltered}{marr}(object)
}
\arguments{
\item{object}{an object of class \code{marr}.}
}
\value{
The percent of reproducible sample pairs based on a
threshold value after applying maximum rank reproducibility.
}
\description{
Given a marr object, this function returns the
marr filtered features

Accessors for the 'marrFeaturesfiltered'
slot of a marr object.
}
\examples{
data <- matrix(rnorm(2400), nrow=200, ncol=12)
data_marr <- marr(object = data, pSamplepairs=0.75,
pFeatures=0.75, alpha=0.05)
marrFeaturesfiltered(data_marr)

}
