% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllslotAccessors.R
\docType{methods}
\name{marrSamplepairs}
\alias{marrSamplepairs}
\alias{marrSamplepairs,marr-method}
\title{Generic function that returns the marr sample pairs}
\usage{
marrSamplepairs(object)

\S4method{marrSamplepairs}{marr}(object)
}
\arguments{
\item{object}{an object of class \code{marr}.}
}
\value{
The distribution of percent reproducible features
(column-wise) per sample pair after applying the maximum
rank reproducibility.
}
\description{
Given a marr object, this function returns the
marr sample pairs

Accessors for the 'marrSamplepairs' slot of
a marr object.
}
\examples{
data <- matrix(rnorm(2400), nrow=200, ncol=12)
data_marr <- marr(object = data, pSamplepairs=0.75,
pFeatures=0.75, alpha=0.05)
marrSamplepairs(data_marr)

}
